Enforce read-only volumes
After further consideration, the warning logs and flag for backcompat
feel a little ineffective. The logs are not a very actionable place to
do deprecation warnings. They are just tucked away too deep for anyone
to notice.

Further, as this will be the only change between 0.1.0 and 0.2.0, so if
enforcement is problematic, operators can just not upgrade (or downgrade
if they don't read the changelog before upgrading).

Signed-off-by: Andrew Harding <aharding@vmware.com>